add new classes
